% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{cells_threshold}
\alias{cells_threshold}
\title{Cells Threshold}
\usage{
cells_threshold(data)
}
\arguments{
\item{data}{A data frame of gene counts.}
}
\value{
A list of thresholds at various percentiles.
}
\description{
This function calculates thresholds for the number of expressed genes per cell.
}
\examples{
cells_threshold(data_frame)
}
