% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{count_genes}
\alias{count_genes}
\title{Count Genes}
\usage{
count_genes(df, count = "counts")
}
\arguments{
\item{df}{A data frame or matrix of gene expression counts, where rows represent genes and columns represent samples or cells.}

\item{count}{A character string specifying the counting method. Acceptable values are:
\describe{
  \item{'counts'}{Calculate the total counts for each sample or cell (default).}
  \item{'genes'}{Count the number of expressed genes (non-zero counts) for each sample or cell.}
}}
}
\value{
A data frame with two columns:
  \describe{
    \item{set_name}{The names of the samples or cells.}
    \item{n}{The total counts or the number of expressed genes, depending on the `count` parameter.}
  }

A data frame with the count of expressed genes per sample.
}
\description{
This function calculates the number of expressed genes per sample / cell.
}
\examples{
count_genes(df, count = 'counts')
}
