% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{features_perc_plot}
\alias{features_perc_plot}
\title{Features percent per Cell / Set Plot}
\usage{
features_perc_plot(data, count = "counts", cut_point = 0)
}
\arguments{
\item{data}{A data frame with percentage of features / features counts..}

\item{cut_point}{Optional cutoff point for visualizing the threshold. Default is 0.}
}
\value{
A ggplot object.
}
\description{
This function generates a plot of the percent of features / features counts per cell /set.
}
\examples{
features_perc_plot(data, count = 'counts', cut_point = 0)
}
