% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{features_percentage}
\alias{features_percentage}
\title{Percent Features}
\usage{
features_percentage(data, features_list)
}
\arguments{
\item{df}{A data frame of gene counts obtained from load_rnaseq()}
}
\value{
A data frame with the percent features / features counts.
}
\description{
This function calculates the percentage of features / feature counts compared to the remaining features / feature counts in a set or cell.
}
\examples{
features_percentage(data, features_list)
}
