% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{genes_per_cell_plot}
\alias{genes_per_cell_plot}
\title{Genes per Cell / Set Plot}
\usage{
genes_per_cell_plot(data, cut_point = NaN)
}
\arguments{
\item{data}{A data frame of gene counts.}

\item{cut_point}{Optional cutoff point for visualizing the threshold. Default is NaN. If NaN, the threshold is set on median.}
}
\value{
A ggplot object.
}
\description{
This function generates a plot of the number of expressed genes per cell.
}
\examples{
genes_per_set_plot(data_frame)
}
