% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{get_cluster_stats}
\alias{get_cluster_stats}
\title{Generate cluster statistics}
\usage{
get_cluster_stats(data, clusters)
}
\arguments{
\item{data}{A matrix or data frame of gene expression data.}

\item{clusters}{A data frame of cluster assignments for each sample.}
}
\value{
A data frame of genes with log-fold change and adjusted p-values for each cluster.
}
\description{
Computes differential expression statistics for genes in each cluster.
}
\examples{
cluster_stats <- get_cluster_stats(data, clusters)
}
