% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{get_clusters}
\alias{get_clusters}
\title{Retrieve clusters from UMAP and DBSCAN results}
\usage{
get_clusters(
  pca_data,
  pc = 10,
  eps = 0.5,
  min_dist = 0.01,
  n_neighbors = 5,
  minPts = 5
)
}
\arguments{
\item{pca_data}{An object of class 'prcomp' containing PCA results.}

\item{pc}{Number of principal components to use for UMAP. Default is 10.}

\item{eps}{The epsilon parameter for DBSCAN. Default is 0.5.}

\item{min_dist}{Minimum distance for UMAP. Default is 0.01.}

\item{n_neighbors}{Number of neighbors for UMAP. Default is 5.}

\item{minPts}{Minimum points for DBSCAN. Default is 5.}
}
\value{
A data frame mapping sample names to cluster assignments.
}
\description{
Extracts cluster assignments for each sample based on UMAP and DBSCAN results.
}
\examples{
clusters <- get_clusters(pca_result)
}
