% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{get_knee}
\alias{get_knee}
\title{Create a knee plot for PCA results}
\usage{
get_knee(pca_data)
}
\arguments{
\item{pca_data}{An object of class 'prcomp' containing PCA results.}
}
\value{
A ggplot2 object showing the knee plot.
}
\description{
Visualizes the explained variance for each principal component to identify the optimal number of components.
}
\examples{
knee_plot <- get_knee(pca_result)
}
