% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{get_PCA}
\alias{get_PCA}
\title{Perform PCA on gene expression data}
\usage{
get_PCA(data)
}
\arguments{
\item{data}{A matrix or data frame of gene expression data.}
}
\value{
An object of class 'prcomp' containing PCA results.
}
\description{
Applies Principal Component Analysis (PCA) to identify the main components of variation in the data.
}
\examples{
pca_result <- get_PCA(data)
}
