% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{heatmap_plot}
\alias{heatmap_plot}
\title{Plot heatmap for selected genes}
\usage{
heatmap_plot(display_data, features, features_metadata = NaN, min_value = 0.5)
}
\arguments{
\item{display_data}{A matrix or data frame of normalized gene expression data.}

\item{genes}{A vector of gene names to include in the heatmap.}
}
\value{
A heatmap visualization of the selected genes.
}
\description{
Creates a heatmap for a subset of genes based on their expression values.
}
\examples{
heatmap_plot <- heatmap_plot(display_data, selected_genes)
}
