% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{load_rnaseq}
\alias{load_rnaseq}
\title{Load scRNA-seq / RNA-seq Data}
\usage{
load_rnaseq(
  path,
  concatenation_type = "sum",
  annotation_sides = c("three_prime_UTR", "exon", "five_prime_UTR")
)
}
\arguments{
\item{path}{Path to the directory containing scRNA-seq / RNA-seq matrices.}

\item{concatenation_type}{Method to concatenate data: 'sum', 'avr', or 'top_diff'.}

\item{annotation_sides}{A vector of annotation sides to include (e.g., 'three_prime_UTR', 'exon', 'five_prime_UTR').}
}
\value{
A list of data frames, each corresponding to an annotation type.
}
\description{
This function processes multiple scRNA-seq / RNA-seq files, filters based on annotation types, and concatenates data as specified.
}
\examples{
load_sc_rnaseq("data/matrices/", concatenation_type = 'sum')
}
