% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{load_single_file}
\alias{load_single_file}
\title{Load Single File from featureCounts output}
\usage{
load_single_file(path)
}
\arguments{
\item{path}{A string specifying the path to the input file.}
}
\value{
A data frame containing the processed data.
}
\description{
This function loads a single scRNA-seq / RNA-seq data file, removes duplicate rows, and renames the 7th column to 'counts'.
}
\examples{
load_single_file("data/matrix.txt")
}
