% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{pca_plot}
\alias{pca_plot}
\title{Create a PCA scatter plot}
\usage{
pca_plot(pca_data)
}
\arguments{
\item{pca_data}{An object of class 'prcomp' containing PCA results.}
}
\value{
A ggplot2 scatter plot of the first two principal components.
}
\description{
Plots the first two principal components from PCA results.
}
\examples{
pca_scatter <- pca_plot(pca_result)
}
