% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{select_on_features}
\alias{select_on_features}
\title{Select Cells Based on Feature Percentage Data}
\usage{
select_on_features(data, features_perc, count = "counts", threshold = NaN)
}
\arguments{
\item{data}{A matrix or data frame of gene expression data, where rows represent genes 
and columns represent cells.}

\item{features_perc}{A data frame containing feature percentage data, including a column 
`set_name` for cell identifiers and `perc_features_counts` or `perc_features` for percentage values.}

\item{count}{A character string specifying the type of data to filter by. Acceptable values 
are 'counts' (for filtering based on feature counts) or 'genes' (for filtering based on 
gene occurrence percentages). Default is 'counts'.}

\item{threshold}{A numeric value specifying the maximum percentage threshold for a cell 
to be retained. Only cells with a percentage below or equal to this threshold will be 
included. If no threshold is provided (default is NaN), the function will print a message 
and perform no operation.}
}
\value{
A filtered dataset (matrix or data frame) containing only the cells that meet 
  the specified threshold criteria. If an invalid `counts` parameter is provided or 
  no threshold is specified, a message will be printed instead.
}
\description{
This function filters cells in a dataset based on the percentage occurrence of features 
or feature counts calculated using `features_percentage()`. It retains only the cells 
that meet a specified threshold.
}
\examples{
filtered_data <- select_on_features(data = reduced_data, features_perc, counts = 'counts', threshold = 20)
}
