% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{umap_db_scan}
\alias{umap_db_scan}
\title{Perform UMAP and DBSCAN clustering}
\usage{
umap_db_scan(
  pca_data,
  pc = 10,
  eps = 0.5,
  min_dist = 0.01,
  n_neighbors = 5,
  minPts = 5
)
}
\arguments{
\item{pca_data}{An object of class 'prcomp' containing PCA results.}

\item{pc}{Number of principal components to use for UMAP. Default is 10.}

\item{eps}{The epsilon parameter for DBSCAN. Default is 0.5.}

\item{min_dist}{Minimum distance for UMAP. Default is 0.01.}

\item{n_neighbors}{Number of neighbors for UMAP. Default is 5.}

\item{minPts}{Minimum points for DBSCAN. Default is 5.}
}
\value{
A ggplot2 UMAP plot with DBSCAN clusters.
}
\description{
Combines UMAP dimensionality reduction with DBSCAN clustering for gene expression data.
}
\examples{
umap_plot <- umap_db_scan(pca_result)
}
