% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{var_plot}
\alias{var_plot}
\title{Plot variable genes}
\usage{
var_plot(var_data, side = c("equal", "variable"), n_top = 20)
}
\arguments{
\item{var_data}{A data frame of variance and mean values for genes.}

\item{side}{Character vector specifying the type of genes to plot ('equal', 'variable'). Default includes both.}

\item{n_top}{Integer specifying the number of top genes to label. Default is 20.}
}
\value{
A ggplot2 scatter plot object.
}
\description{
Generates a scatter plot of gene variance and mean, highlighting the most variable genes.
}
\examples{
var_gene_plot <- var_plot(var_data)
}
